+++ datumaro/cli/commands/merge.py
@@ -10,13 +10,15 @@ import os
 import os.path as osp
 
 from datumaro.components.dataset import DEFAULT_FORMAT
+from datumaro.components.environment import Environment
 from datumaro.components.errors import (
     DatasetMergeError, DatasetQualityError, ProjectNotFoundError,
 )
 from datumaro.components.operations import IntersectMerge
+from datumaro.components.project import ProjectBuildTargets
 from datumaro.util.scope import scope_add, scoped
 
-from ..util import MultilineFormatter
+from ..util import MultilineFormatter, join_cli_args
 from ..util.errors import CliException
 from ..util.project import (
     generate_next_file_name, load_project, parse_full_revpath,
@@ -26,12 +28,16 @@ from ..util.project import (
 def build_parser(parser_ctor=argparse.ArgumentParser):
     parser = parser_ctor(help="Merge few projects",
         description="""
-        Merges multiple datasets into one and produces a new
-        dataset in the default format. This can be useful if you
-        have few annotations and wish to merge them,
-        taking into consideration potential overlaps and conflicts.
-        This command can try to find a common ground by voting or
+        Merges multiple datasets into one and produces a new dataset.
+        The command can be useful if you have few annotations and wish
+        to merge them, taking into consideration potential overlaps and
+        conflicts. This command can try to find common ground by voting or
         return a list of conflicts.|n
+        |n
+        In simple cases, when dataset images do not intersect and new
+        labels are not added, the recommended way of merging is using the
+        "patch" command. It will offer better performance and provide the same
+        results.|n
         |n
         This command has multiple forms:|n
         1) %(prog)s <revpath>|n
@@ -56,6 +62,13 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         It can be useful for dataset paths in targets. When not specified,
         the current project's working tree is used.|n
         |n
+        The output format can be specified with the '-f/--format' option.
+        Each dataset format has its own export
+        options, which are passed after the '--' separator (see examples),
+        pass '-- -h' for more info. If not stated otherwise, by default
+        only annotations are exported; to include images pass
+        '--save-images' parameter.|n
+        |n
         Examples:|n
         - Merge annotations from 3 (or more) annotators:|n
         |s|s%(prog)s project1/ project2/ project3/|n
@@ -72,12 +85,17 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         |n
         - Merge a source from a previous revision and a dataset:|n
         |s|s%(prog)s HEAD~2:source-2 path/to/dataset2:yolo
+        |n
+        - Merge datasets and save in different format:|n
+        |s|s%(prog)s -f voc dataset1/:yolo path2/:coco -- --save-images
         """,
         formatter_class=MultilineFormatter)
 
     def _group(s):
         return s.split(',')
 
+    parser.add_argument('_positionals', nargs=argparse.REMAINDER,
+        help=argparse.SUPPRESS) # workaround for -- eaten by positionals
     parser.add_argument('targets', nargs='+',
         help="Target dataset revpaths (repeatable)")
     parser.add_argument('-iou', '--iou-thresh', default=0.25, type=float,
@@ -97,8 +115,14 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         help="Output directory (default: generate a new one)")
     parser.add_argument('--overwrite', action='store_true',
         help="Overwrite existing files in the save directory")
+    parser.add_argument('-f', '--format', default=DEFAULT_FORMAT,
+        help="Output format (default: %(default)s)")
     parser.add_argument('-p', '--project', dest='project_dir',
         help="Directory of the 'current' project (default: current dir)")
+    parser.add_argument('extra_args', nargs=argparse.REMAINDER,
+        help="Additional arguments for converter (pass '-- -h' for help). "
+            "Must be specified after the main command arguments and after "
+            "the '--' separator")
     parser.set_defaults(command=merge_command)
 
     return parser
@@ -110,6 +134,22 @@ def get_sensitive_args():
 
 @scoped
 def merge_command(args):
+    # Workaround. Required positionals consume positionals from the end
+    args._positionals += join_cli_args(args, 'targets', 'extra_args')
+
+    has_sep = '--' in args._positionals
+    if has_sep:
+        pos = args._positionals.index('--')
+        if pos == 0:
+            raise argparse.ArgumentError(None,
+                message="Expected at least 1 target argument")
+    else:
+        pos = len(args._positionals)
+    args.targets = args._positionals[:pos] or [ProjectBuildTargets.MAIN_TARGET]
+    args.extra_args = args._positionals[pos + has_sep:]
+
+    show_plugin_help = '-h' in args.extra_args or '--help' in args.extra_args
+
     dst_dir = args.dst_dir
     if dst_dir:
         if not args.overwrite and osp.isdir(dst_dir) and os.listdir(dst_dir):
@@ -123,8 +163,21 @@ def merge_command(args):
     try:
         project = scope_add(load_project(args.project_dir))
     except ProjectNotFoundError:
-        if args.project_dir:
+        if not show_plugin_help and len(args.targets) == 1 and args.project_dir:
             raise
+
+    if project is not None:
+        env = project.env
+    else:
+        env = Environment()
+
+    try:
+        converter = env.converters[args.format]
+    except KeyError:
+        raise CliException("Converter for format '%s' is not found" % \
+            args.format)
+
+    export_args = converter.parse_cmdline(args.extra_args)
 
     source_datasets = []
     try:
@@ -144,7 +197,8 @@ def merge_command(args):
         output_conf_thresh=args.output_conf_thresh, quorum=args.quorum
     ))
     merged_dataset = merger(source_datasets)
-    merged_dataset.export(save_dir=dst_dir, format=DEFAULT_FORMAT)
+
+    merged_dataset.export(save_dir=dst_dir, format=converter, **export_args)
 
     report_path = osp.join(dst_dir, 'merge_report.json')
     save_merge_report(merger, report_path)