+++ datumaro/plugins/lfw_format.py
@@ -13,6 +13,7 @@ from datumaro.components.converter import Converter
 from datumaro.components.extractor import DatasetItem, Importer, SourceExtractor
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.util.image import find_images
+from datumaro.util.meta_file_util import has_meta_file, parse_meta_file
 
 
 class LfwPath:
@@ -44,6 +45,10 @@ class LfwExtractor(SourceExtractor):
         self._items = list(self._load_items(path).values())
 
     def _load_categories(self, path):
+        if has_meta_file(self._dataset_dir):
+            return { AnnotationType.label: LabelCategories().
+                from_iterable(list(parse_meta_file(self._dataset_dir).keys())) }
+
         label_cat = LabelCategories()
         if osp.isfile(path):
             with open(path, encoding='utf-8') as labels_file:
@@ -188,6 +193,10 @@ class LfwConverter(Converter):
     DEFAULT_IMAGE_EXT = LfwPath.IMAGE_EXT
 
     def apply(self):
+        os.makedirs(self._save_dir, exist_ok=True)
+        if self._save_dataset_meta:
+            self._save_meta_file(self._save_dir)
+
         for subset_name, subset in self._extractor.subsets().items():
             label_categories = self._extractor.categories()[AnnotationType.label]
             labels = {label.name: 0 for label in label_categories}